+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed oracle
 
 ## Added
+- Error message `Invalid Base Cnrcy`
 - `upper/lower_slippage_bound` for `BuyMarketSeat` and `SellMarketSeat` instruction data
 
 ## [v2.9] - 2025-11-18